brief 154: double-precision the checkpointed float-trunc CP9 F/B kernels
Brief 153 root-caused the genome-scale `-g` truncated band collapse (HSV/MPXV
"no valid parsetree") as float32 accumulation roundoff in the CP9 Forward/
Backward recursion: ~0.24 ulp/row accumulating linearly in L (-1.69 nats at
norovirus L=7538, ~-35 at HSV, ~-45 at MPXV), deflating posteriors to ~0 so
bands prune to nothing. A double-precision F/B mirror drove the norovirus F/B
gap from -1.687 to +0.00003 with a flat per-row marginal.

This converts ONLY the checkpointed float-trunc family (the `cp9_chk_*F` kernels,
cp9chkF_t store/fills, cp9segF_t segment/posterior/dual-sweep reduction, and
cp9_PredictStartAndEndFromPoccF) in cm_p7_band_chk.c to double, end to end
(cells + accumulation + checkpoint storage + streamed pocc_arr), following
brief 153's mechanical float->double transform. The non-checkpointed
cp9_*P7BF float kernels and the entire INT (non-trunc) section are untouched,
which gates the fix by codepath: genome-scale truncated runs require
--p7ibv-ckpt anyway (the non-ckpt F/B hits the ~33.6 GB floor), so the working
genome path gets double; sub-genome trunc stays float-fast on the non-ckpt path.

Details:
- p7_FLogsum (float LUT) -> cp9_chk_dlogsum (exact log1p/exp, double-valued).
  153 showed exact-vs-LUT doesn't move the gap, but a double-valued logsum is
  required here regardless.
- The 5 row kernels, cp9chkF_t (+Create/FwdFill/BwdFill), cp9segF_t
  (+Create/Fill/PostRow), cp9_FB2HMMBandsP7BF_chk (mass/sc/thresh/pocc all
  double; esl_vec_FSet->DSet; expf->exp; logf->log), and the predictor: double.
- pocc_arr widened to double in cp9_FB2HMMBandsP7BF_chk (def + infernal.h decl)
  and cp9_FBMatrices2BandsP7BF_chk's allocation. The public entry point
  cp9_FBMatrices2BandsP7BF_chk signature is unchanged.
- Checkpointing keeps the double F/B at O(sqrt(L)) memory: ~2x of ~58 MB at HSV.
- Added a file-local CP9_DMX double matrix (CreateCP9DMatrix/FreeCP9DMatrix) used
  only by the env-gated CP9_CKPTF_FBCMP debug block, whose chk-vs-nonchk compare
  is now float-ref-vs-double-chk with tolerance.

+++ src/infernal.h
@@ -3276,7 +3276,7 @@ extern int          cp9_FB2HMMBandsP7B(CP9_t *hmm, char *errbuf, ESL_DSQ *dsq, C
 extern int          cp9_FB2HMMBandsP7B_chk(CP9_t *hmm, char *errbuf, ESL_DSQ *dsq, CP9Bands_t *cp9b, int L, int M, double p_thresh, int *kmin, int *kmax, int debug_level, int do_pnmono, int do_pnmono_print);
 extern int          cp9_FBMatrices2BandsP7B_chk(CM_t *cm, char *errbuf, CP9_t *cp9, ESL_DSQ *dsq, CP9Bands_t *cp9b, int *kmin, int *kmax, int L, int i0, int j0, int pass_idx, int debug_level, int do_pnmono, int do_pnmono_print);
 /* checkpointed banded CP9 P7B float/truncated F/B (brief 150, Phase 2, cm_p7_band_chk.c) */
-extern int          cp9_FB2HMMBandsP7BF_chk(CP9_t *hmm, char *errbuf, ESL_DSQ *dsq, CP9Bands_t *cp9b, int L, int M, double p_thresh, int *kmin, int *kmax, int debug_level, int do_pnmono, int do_pnmono_print, float *pocc_arr);
+extern int          cp9_FB2HMMBandsP7BF_chk(CP9_t *hmm, char *errbuf, ESL_DSQ *dsq, CP9Bands_t *cp9b, int L, int M, double p_thresh, int *kmin, int *kmax, int debug_level, int do_pnmono, int do_pnmono_print, double *pocc_arr);
 extern int          cp9_FBMatrices2BandsP7BF_chk(CM_t *cm, char *errbuf, CP9_t *cp9, ESL_DSQ *dsq, CP9Bands_t *cp9b, int *kmin, int *kmax, int L, int i0, int j0, int pass_idx, int debug_level, int do_pnmono, int do_pnmono_print);
 extern int          p7_Seq2Bands(CM_t *cm, char *errbuf, P7_PROFILE *gm, P7_GMX *gx, P7_BG *bg, P7_TRACE *p7_tr, ESL_DSQ *dsq, int L,
 				 double **phi, float sc7, int len7, int end7, float mprob7, float mcprob7, float iprob7, float ilprob7, int pad7,
